refactor(dev): promote dependencies module to root of remix-dev
